+++ Changelog.md
@@ -3,6 +3,8 @@
 ## 0.0.2 - release
 
 ### Feature
+ - add compare method for `ParseOptions` and `Job` struct
+ - add reset function for task manager
  - add `version` command
  - add flag for log level [(#5)](https://github.com/AmolKumarGupta/crona/issues/5)
  - add config flag to command [(#3)](https://github.com/AmolKumarGupta/crona/issues/3)